+++ CHANGELOG.md
@@ -35,6 +35,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Sometimes keyboard table is duplicating shortcuts
 - Export Project Database task not exporting rows whose `project_id` is `NULL` [#3203]
 - Close icon is difficult to distinguish when modal background is transparent [#3245]
+- Missing identifiers and determinations on collection object table in New collecting event task [#3246]
 
 ### Changed
 - Lock, navigation, UI, and code refreshments to Simple new specimen [#3190]
@@ -63,6 +64,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 [#3234]: https://github.com/SpeciesFileGroup/taxonworks/issues/3234
 [#3238]: https://github.com/SpeciesFileGroup/taxonworks/issues/3238
 [#3245]: https://github.com/SpeciesFileGroup/taxonworks/issues/3245
+[#3246]: https://github.com/SpeciesFileGroup/taxonworks/issues/3246
 
 ## [0.29.6] - 2022-11-08